+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Infer span descriptions via `sentry-conventions`. ([#6093](https://github.com/getsentry/relay/pull/6093))
+- Raises the size limit for the flags context to 64KiB. ([#6137](https://github.com/getsentry/relay/pull/6137))
 
 **Bug Fixes**: